refactor: TableWithFilters 모듈화 (#77)
